% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_np.R
\name{calculate_np}
\alias{calculate_np}
\title{Calculate Number of Patches (NP) in a Raster}
\usage{
calculate_np(raster)
}
\arguments{
\item{raster}{A \code{SpatRaster} object (from the \code{terra} package) representing
the landscape. Each unique value corresponds to a land-cover class.}
}
\value{
tibble (data frame) containing the number of patches for each class.
}
\description{
This function calculates the \strong{number of patches (NP)} in a landscape raster
using 8-neighbor connectivity. It wraps the \code{lsm_c_np} function from the
\code{landscapemetrics} package for easier use.
}
\examples{
library(terra)
 # create a small example raster
r <- rast(nrows = 10, ncols = 10)
values(r) <- sample(1:3, 100, replace = TRUE)
calculate_np(r)

}
