% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_forest.R
\name{preprocess_forest}
\alias{preprocess_forest}
\title{Preprocess Forest Raster}
\usage{
preprocess_forest(raster, forest_class = "trees")
}
\arguments{
\item{raster}{SpatRaster. Land cover raster.}

\item{forest_class}{Integer. Value representing forest class.}
}
\value{
SpatRaster (binary: 1 = forest, 0 = non-forest)
}
\description{
Clips land cover raster to AOI and creates a binary forest/non-forest raster.
}
\examples{
\dontrun{
binary <- preprocess_forest(raster, aoi, forest_class = 10)
}

}
