% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_and_project.R
\name{validate_and_project}
\alias{validate_and_project}
\title{Validate and Project Spatial Data}
\usage{
validate_and_project(raster, aoi)
}
\arguments{
\item{raster}{A SpatRaster object}

\item{aoi}{A SpatVector (polygon) representing the study area.}
}
\value{
A list containing:
\itemize{
\item raster: Projected raster
\item aoi: Projected AOI
}
}
\description{
Ensures that the input raster and AOI share the same coordinate reference
system (CRS) and are in a projected CRS suitable for spatial analysis.
If the CRS is geographic (e.g., WGS84), the data will be reprojected.
}
\examples{
\dontrun{
result <- validate_and_project(raster, aoi)
}


}
